pyproject.toml
[project]
name = "anthropic-courses"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.45.2",
    "python-dotenv>=1.0.1",
]
